好久没提交了，也不知道改了哪些，都提交了吧

+++ src/main/java/com/sdf/core/mapper/demo/DemoUserDao.java
@@ -8,7 +8,7 @@ import java.util.List;
 @Repository
 public interface DemoUserDao {
 
-    DemoUser selectById(long id);
+    DemoUser selectById(int id);
 
     DemoUser selectByName(String user_name);
